image-forgery-swin-pytorch/requirements.txt
torch==1.13.0
torchvision==0.14.0
timm==0.6.12
Pillow==9.2.0
numpy==1.23.1
scikit-learn==1.1.2
matplotlib==3.5.2
tqdm==4.64.0
PyYAML==6.0